# Buffered S3 reader