--- a/pfio/v2/s3.py
+++ b/pfio/v2/s3.py
@@ -9,7 +9,6 @@ from botocore.exceptions import ClientError
 
 from .fs import FS, FileStat
 
-
 DEFAULT_MAX_BUFFER_SIZE = 16 * 1024 * 1024