@kuenishi 2021-12-27
Thank you for very thoughtful suggestion with real benchmarks. I'd welcome the idea and that's what I wanted to do, indicated in the TODO comment.

My suggestion would be add `buffering` keyword argument as well as in `open()` builtin function and like in other FS systems, to align the interface regarding the potential possible options like in `open_url(path, buffering=-1)` . See also [Local#open](https://github.com/pfnet/pfio/blob/dc1998fd460a3ac48834c24d07bc6e3f6324864b/pfio/v2/local.py#L63-L70) and Hdfs#open . [The document](https://docs.python.org/3/library/functions.html#open) of `open()` says:

> buffering is an optional integer used to set the buffering policy. Pass 0 to switch buffering off (only allowed in binary mode), 1 to select line buffering (only usable in text mode), and an integer > 1 to indicate the size in bytes of a fixed-size chunk buffer.

Thus the default value of 32MB would be nice, but I want to leave a mean for app developers to configure the buffer size.